move displaDbIliMode to plugin globals, move DbActionType to libqgsprojectgen lib global
